# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_emzola_issuetracker_internal_model.User:
    properties:
      activated:
        type: boolean
      created_by:
        type: string
      created_on:
        type: string
      email:
        type: string
      id:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  internal_handler_http.activateUserPayload:
    properties:
      token:
        type: string
    type: object
  internal_handler_http.assignUserToProjectPayload:
    properties:
      project_id:
        type: integer
    type: object
...
  internal_handler_http.createUserPayload:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
...
  internal_handler_http.updateUserPayload:
    properties:
      email:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
host: https://issuetracker-api-dev.fl0.io
info:
  contact:
    email: emma.idika@yahoo.com
    name: API Support
  description: This is an API service for an issue tracker.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Issue Tracker API
  version: 1.0.0
paths:
...
  /v1/users:
    get:
      description: This endpoint gets all users
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Query string param for name
        in: query
        name: name
        type: string
      - description: Query string param for email
        in: query
        name: email
        type: string
      - description: Query string param for role
        in: query
        name: role
        type: string
      - description: Query string param for pagination (min 1)
        in: query
        name: page
        type: string
      - description: Query string param for pagination (max 100)
        in: query
        name: page_size
        type: string
      - description: 'Sort by asc or desc order. Asc: id, name, email, created_on,
          modified_on | Desc: -id, -name, -email, -created_on, -modified_on'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.User'
            type: array
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with the request payload
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Request payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handler_http.createUserPayload'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.User'
        "400":
          description: Bad Request
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Create a new user
      tags:
      - users
  /v1/users/{user_id}:
    delete:
      description: This endpoint deletes a user
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: ID of user to delete
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a user
      tags:
      - users
    get:
      description: This endpoint gets a user by ID
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: ID of user to get
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.User'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: This endpoint updates a user
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Request payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handler_http.updateUserPayload'
      - description: ID of user to update
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.User'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "409":
          description: Conflict
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Update a user
      tags:
      - users
  /v1/users/{user_id}/projects:
    get:
      description: This endpoint gets all projects for a user
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Query string param for pagination (min 1)
        in: query
        name: page
        type: string
      - description: Query string param for pagination (max 100)
        in: query
        name: page_size
        type: string
      - description: 'Sort by asc or desc order. Asc: id | Desc: -id'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.User'
            type: array
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Get all projects for user
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Assign a user to a project with the request payload
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Request payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handler_http.assignUserToProjectPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Assign a user to a project
      tags:
      - users
  /v1/users/activated:
    put:
      consumes:
      - application/json
      description: Activate a new user with the request payload
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Request payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handler_http.activateUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.User'
        "400":
          description: Bad Request
        "409":
          description: Conflict
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Activate a new user
      tags:
      - users
swagger: "2.0"